requirements.txt
awsebcli==3.14.8
Babel==2.6.0
blessed==1.15.0
botocore==1.12.75
cached-property==1.5.1
cement==2.8.2
certifi==2018.10.15
chardet==3.0.4
colorama==0.3.9
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.1.2
django-allauth==0.38.0
django-appconf==1.0.2
django-extensions==2.1.4
django-filter==2.0.0
django-ipware==2.1.0
django-rest-auth==0.9.3
django-scheduler==0.8.8
djangorestframework==3.8.2
djangorestframework-filters==1.0.0.dev0
docker==3.6.0
docker-compose==1.21.2
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
gunicorn==19.9.0
icalendar==4.0.3
idna==2.6
itypes==1.1.0
Jinja2==2.10
jmespath==0.9.3
jsonfield==2.0.2
jsonschema==2.6.0
MarkupSafe==1.0
oauthlib==2.1.0
pathspec==0.5.5
phonenumbers==8.10.1
Pillow==5.3.0
pinax-eventlog==2.0.3
pinax-stripe==4.4.0
psycopg2==2.7.6.1
pygraphviz==1.5
python-dateutil==2.7.5
python3-openid==3.1.0
pytz==2018.5
PyYAML==3.13
requests==2.18.4
requests-oauthlib==1.0.0
semantic-version==2.5.0
six==1.11.0
stripe==2.16.0
termcolor==1.1.0
texttable==0.9.1
uritemplate==3.0.0
urllib3==1.22
wcwidth==0.1.7
websocket-client==0.54.0
whitenoise==4.1.2
yamllint==1.13.0
